added text collector

+++ main_generator.py
@@ -1,4 +1,5 @@
 from rich import print
+from rich.console import Console
 import glob
 import os
 from spacy.lang.de import German
@@ -18,6 +19,8 @@ import numpy as np
 import webrtcvad
 from scipy.ndimage.morphology import binary_dilation
 import soundfile
+import sys
+
 
 chunk = 1024
 sample_format = pyaudio.paInt16
@@ -70,14 +73,18 @@ def trim_long_silences(wav):
 
 
 if __name__ == '__main__':
+
+	console = Console()
+	
 	table = Table()
 	table.add_column("TSS Dataset Creator", style="cyan")
 	table.add_row("2021 - padmalcom")
 	table.add_row("www.stonedrum.de")
 	
-	print(table)
-
-	print("\nPlease select your [red]microphone[/red] (enter the device number).")
+	
+	console.print(table)
+	
+	console.print("\nPlease select your [red]microphone[/red] (enter the device number).")
 	# Initialisiere pyaudio
 	pyaudio = pyaudio.PyAudio()
 	
@@ -88,35 +95,45 @@ if __name__ == '__main__':
 		if (pyaudio.get_device_info_by_host_api_device_index(0, i).get('maxInputChannels')) > 0:
 			print ("Input Device id ", i, " - ", pyaudio.get_device_info_by_host_api_device_index(0, i).get('name'))
 	in_mic_id = int(input())
-	print("You have selected [red]%s[/red] as input device." % pyaudio.get_device_info_by_host_api_device_index(0, in_mic_id).get('name'))
+	console.print("You have selected [red]%s[/red] as input device." % pyaudio.get_device_info_by_host_api_device_index(0, in_mic_id).get('name'))
 	
 	
 	# 0. select folder to save wavs
 	app_folder = os.path.dirname(os.path.realpath(__file__))
 	project_folder = os.path.join(app_folder, "sample_" + datetime.now().strftime("%d.%m.%Y"))
-	print("Please select a [red]folder[/red] to save your current session to (default [i]%s[/i])." % project_folder)
+	console.print("Please select a [red]folder[/red] to save your current session to (default [i]%s[/i])." % project_folder)
 	in_folder = input()
 	if not in_folder:
 		in_folder = project_folder
 	if not os.path.exists(in_folder):
 		os.mkdir(in_folder)
-	print("wavs and transcripts will be saved in [red]%s[/red]" % in_folder)
+	console.print("wavs and transcripts will be saved in [red]%s[/red]" % in_folder)
 	
 	
 	# 0.1 select languages
-	print("Please select a [red]language[/red] (default [i]%s[/i])." % 'de')
+	console.print("Please select a [red]language[/red] (default [i]%s[/i])." % 'de')
 	in_lang = input()
 	if not in_lang:
 		in_lang = 'de'
-	print("Language is set to [red]%s[/red]." % in_lang)
+	console.print("Language is set to [red]%s[/red]." % in_lang)
 	
+	# Save to csv or txt?
+	console.print("Save to [red]csv[/red] or to single [red]txt[/red] file(s)?  (default [i]csv[/i]).")
+	in_text_format = input()
+	if not in_text_format:
+		in_text_format = 'csv'
+	if not in_text_format == 'txt':
+		in_text_format = 'csv'
+	console.print("Format set to [red]%s[/red]." % in_text_format)
+	
+	# Display files found
 	lang_text_files = glob.glob(os.path.join(app_folder, 'texts', in_lang + '*.txt'))
-	print("Found %d text files for %s" % (len(lang_text_files), in_lang))
+	console.print("Found %d text files for %s" % (len(lang_text_files), in_lang))
 	
 	if len(lang_text_files) == 0:
-		print("Please select another language or create text files starting with %s in the [red]texts[/red] folder. Exiting." % in_lang)
+		console.print("Please select another language or create text files starting with %s in the [red]texts[/red] folder. Exiting." % in_lang)
 		sys.exit(0)
-		
+				
 	# 1. Show text, sentence by sentence
 	all_texts = ''
 	for tf in lang_text_files:
@@ -131,7 +148,7 @@ if __name__ == '__main__':
 	elif in_lang == 'fr':
 		nlp = French()
 	else:
-		print("The language %s is not supported yet. Please create a github issue." % in_lang)
+		console.print("The language %s is not supported yet. Please create a github issue." % in_lang)
 		sys.exit(0)
 	
 	nlp.add_pipe('sentencizer')
@@ -139,19 +156,24 @@ if __name__ == '__main__':
 	doc = nlp(all_texts)
 	all_sentences = [str(sent).strip() for sent in doc.sents]
 
-	print("Found %d sentences." % len(all_sentences))
+	console.print("Found %d sentences." % len(all_sentences))
 
 	if len(all_sentences) == 0:
-		print("You need to add some sentences to your text files first. Exiting.")
+		console.print("You need to add some sentences to your text files first. Exiting.")
 		sys.exit(0)
 	
 	# 1.1 Hit Space to skip to save+next, hit backspace to discard
-	print("[green]n[/green] = next sentence, [yellow]d[/yellow] = discard and repeat last recording, [red]e[/red] = exit recording.")
+	console.print("[green]n[/green] = next sentence, [yellow]d[/yellow] = discard and repeat last recording, [red]e[/red] = exit recording.")
+	
+	console.print("Ready?", style="green")
+	input("Press any key to start")
+	
 	i = 0
 	cancelled = False
 	while i < len(all_sentences):
-		print("[magenta]" + all_sentences[i] + "[/magenta]")
-		print("(%d/%d) [green]n[/green] = next sentence, [yellow]d[/yellow] = discard and repeat last recording, [red]e[/red] = exit recording." % ((i+1), len(all_sentences)))
+		console.clear()
+		console.print("\n\n" + all_sentences[i] + "\n\n", style = "black on white", justify="center")
+		console.print("(%d/%d) [green]n[/green] = next sentence, [yellow]d[/yellow] = discard and repeat last recording, [red]e[/red] = exit recording." % ((i+1), len(all_sentences)))
 		
 		start_time = time.time()
 		current_time = time.time()
@@ -175,7 +197,8 @@ if __name__ == '__main__':
 					data = stream.read(chunk)
 					frames.append(data)
 					stream.close()
-					wf = wave.open(os.path.join(project_folder, str(i) + '.wav'), 'wb')
+					wav_file_name = (str(i) + '.wav').rjust(10, '0')
+					wf = wave.open(os.path.join(project_folder, wav_file_name), 'wb')
 					wf.setnchannels(channels)
 					wf.setsampwidth(pyaudio.get_sample_size(sample_format))
 					wf.setframerate(frame_rate)										
@@ -183,18 +206,25 @@ if __name__ == '__main__':
 					wf.close()
 					
 					# trim silence?
-					wav, source_sr = librosa.load(str(os.path.join(project_folder, str(i) + '.wav')), sr=None)
+					wav, source_sr = librosa.load(str(os.path.join(project_folder, wav_file_name)), sr=None)
 					wav = trim_long_silences(wav)
-					soundfile.write(str(os.path.join(project_folder, str(i) + '.wav')), wav, source_sr)
+					soundfile.write(str(os.path.join(project_folder, wav_file_name)), wav, source_sr)
 										
 					# Write the transcript
-					text_file_path = os.path.join(project_folder, str(i) + '.txt')
-					if os.path.exists(text_file_path):
-						os.remove(text_file_path)
-					text_file = open(text_file_path, 'a')
-					text_file.write(all_sentences[i])
-					text_file.close()
-					# save audio
+					if in_text_format == 'txt':
+						text_file_name = (str(i) + '.txt').rjust(10, '0')
+						text_file_path = os.path.join(project_folder, text_file_name)
+						if os.path.exists(text_file_path):
+							os.remove(text_file_path)
+						text_file = open(text_file_path, 'a')
+						text_file.write(all_sentences[i])
+						text_file.close()
+					else:
+						csv_file_path = os.path.join(project_folder, 'metadata.csv')
+						csv_file = open(csv_file_path, 'a')
+						# todo cleanse text
+						csv_file.write(wav_file_name + "," + all_sentences[i] + "," + all_sentences[i] + '\n')
+						csv_file.close()
 					break
 				elif keyboard.is_pressed("d"):
 					stream.close()
@@ -207,9 +237,5 @@ if __name__ == '__main__':
 			if cancelled:
 				break
 		
-	# 1.2 Show remaining sentences
-	# 1.3 Hit Escape to quit
 
-	# 2. Ask to process wavs
-	# 2.1 Remove Pauses
 	pyaudio.terminate()
